+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## [Unreleased]
 
-_No unreleased changes._
+### Added
+- Add `patterns = [...]` policy rule support for expanding several URL patterns with shared rule attributes into singular effective rules.
 
 ## [0.0.12] - 2026-03-19